+++ packages/core/src/actions/accounts/getAccount.ts
@@ -1,14 +1,16 @@
-import type { BaseProvider } from '@ethersproject/providers'
+import { providers } from 'ethers'
 
 import { Client, Data, getClient } from '../../client'
 
-export type GetAccountResult<TProvider extends BaseProvider = BaseProvider> = {
+export type GetAccountResult<
+  TProvider extends providers.BaseProvider = providers.BaseProvider,
+> = {
   address?: Data<TProvider>['account']
   connector?: Client<TProvider>['connector']
 }
 
 export function getAccount<
-  TProvider extends BaseProvider,
+  TProvider extends providers.BaseProvider,
 >(): GetAccountResult<TProvider> {
   const { data, connector } = getClient()
   return {